Moved controllers, models, routes, and views to app subdirectory
