+++ app/views/includes/navbar.jade
