+++ server.js
@@ -26,7 +26,7 @@ function compile(str, path) {
 }
 
 // view engine setup
-app.set('views', path.join(__dirname, 'views'));
+app.set('views', path.join(__dirname, 'app/views'));
 app.set('view engine', 'jade');
 
 // uncomment after placing your favicon in /public
@@ -41,7 +41,7 @@ app.use(stylus.middleware({
 }));
 app.use(express.static(path.join(__dirname, 'public')));
 
-expressPath(app, 'routeMap');
+expressPath(app, 'app/routes', { 'controllersPath' : 'app/controllers' });
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
